+++ Postprocessor-microservice/src/main/java/com/postprocessor/Postprocessormicroservice/PostprocessorController.java
@@ -1,0 +1,14 @@
+package com.postprocessor.Postprocessormicroservice;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class PostprocessorController {
+	
+	@RequestMapping("/")
+	public String index() {
+		return "Greetings from Spring Boot!";
+	}
+
+}
