--- preprocessor-microservice/HELP.md
@@ -1,13 +1,0 @@
-# Read Me First
-The following was discovered as part of building this project:
-
-* The original package name 'com.preprocessor.preprocessor-microservice' is invalid and this project uses 'com.preprocessor.preprocessormicroservice' instead.
-
-# Getting Started
-
-### Reference Documentation
-For further reference, please consider the following sections:
-
-* [Official Apache Maven documentation](https://maven.apache.org/guides/index.html)
-* [Spring Boot Maven Plugin Reference Guide](https://docs.spring.io/spring-boot/docs/2.2.4.RELEASE/maven-plugin/)
-
