--- preprocessor-microservice/src/main/java/com/preprocessor/preprocessormicroservice/PreprocessorController.java
@@ -1,14 +1,0 @@
-package com.preprocessor.preprocessormicroservice;
-
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@RestController
-public class PreprocessorController {
-	
-	@RequestMapping("/")
-		public String index() {
-			return "Greetings from Spring Boot!";
-		}
-	
-}
